Add stats.php for displaying statistics using Chart.js
This commit introduces a new stats.php file which fetches data from a SQLite database and displays it using a bar chart implemented with Chart.js. The statistics include dates and accumulated kilometer values from the 'kiliometerliste' table. This will enable visual representation of the data on the web page.